test: split aggregation unit suites
